func_80047480 MATCH; three parks improved off the permuter-output sweep; 874 -> 875
func_80047480: one moved store (60 -> 9), then the counter-init lever twice
over -- i = 0 and j = 0 both hoisted out of their for clauses.  Each alone
is 6 and 3; together they match.

func_80048D08 -1/59 -> 131/131 and 19 on a borrowed name; func_8004D75C
40 -> 21 on the identical-arms hint; func_8004A518 +1/73 -> 112/112 and 27
on u16, which is a documented false zero -- the andi it adds is not in the
target and only cancels a hoist the s32 version has.

func_8002C6C8's find is rejected: it deletes a store the target has, the
third output of that class.

+++ parked/func_8004A518.c
@@ -7,7 +7,7 @@ void func_8004A518(void) {
     s32 *t;
     s32 i;
     s32 o1;
-    s32 o2;
+    u16 o2;
     s32 o3;
     s32 m;
     s32 acc;
